git-svn-id: http://svn.code.sf.net/p/holdingnuts/code/tags/release-0.0.4@628 00c8e4f2-bc05-11dd-9892-7df5a783e2f7

+++ src/server/GameController.cpp
@@ -340,7 +340,7 @@ void GameController::sendTableSnapshot(Table *t)
 			t->seats[t->dealer].seat_no,
 			t->seats[t->sb].seat_no,
 			t->seats[t->bb].seat_no,
-			t->cur_player == -1 ? -1 : t->seats[t->cur_player].seat_no,
+			(t->cur_player == -1) ? -1 : (int)t->seats[t->cur_player].seat_no,
 			t->seats[t->last_bet_player].seat_no);
 		sturn = tmp;
 	}
